Broke home out into it's own folder for cleanliness, and moved tests for home to their own file to reduce toilet paper documentation.  Removed home tests from coverage
