+++ crates/core/TestSettings.toml
@@ -1,3 +1,3 @@
 [ao3-credentials]
-username="Daffidill23"
-password="f1@mingAdr@g0n"
+username=""
+password=""
